requirements.txt
aniso8601==8.0.0
appdirs==1.4.4
astroid==2.4.2
attrs==19.3.0
cached-property==1.5.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
defusedxml==0.6.0
Deprecated==1.2.10
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
flask-marshmallow==0.13.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
idna==2.10
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
lxml==4.5.1
MarkupSafe==1.1.1
marshmallow==3.6.1
mccabe==0.6.1
mysqlclient==2.0.1
pipenv==2020.6.2
pycep-correios==4.0.3
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.3
pytz==2020.1
requests==2.24.0
requests-toolbelt==0.9.1
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.1
Unidecode==1.1.1
urllib3==1.25.9
virtualenv==20.0.25
virtualenv-clone==0.5.4
Werkzeug==1.0.1
wrapt==1.12.1
zeep==3.4.0
